Rename one file
