+++ pyqt_frameless_window/src/c.py
